Move python scripts to modules
