+++ cmake/Common.cmake
@@ -131,3 +131,18 @@ macro(add_flag_if_available flag)
 endmacro()
 
 include("${CMAKE_CURRENT_LIST_DIR}/TupleTreeGenerator.cmake")
+
+function(check_python_requirements)
+  execute_process(
+    COMMAND pip3 freeze
+    RESULT_VARIABLE RETURNCODE
+    OUTPUT_VARIABLE PYTHON_PACKAGES
+    ERROR_FILE "/dev/null"
+  )
+  foreach(REQUIREMENT ${ARGN})
+    string(REGEX MATCH "\n${REQUIREMENT}" RESULT "\n${PYTHON_PACKAGES}")
+    if("${RESULT}" STREQUAL "")
+      message(FATAL_ERROR "Python requirement missing: ${REQUIREMENT}")
+    endif()
+  endforeach()
+endfunction()
