+++ python/pyproject.toml
@@ -1,0 +1,7 @@
+[build-system]
+requires = [
+    "setuptools",
+    "wheel",
+    "datamodel-code-generator"
+]
+build-backend = "setuptools.build_meta"
