+++ python/revng/__init__.py
@@ -1,6 +1,3 @@
 #
 # This file is distributed under the MIT License. See LICENSE.md for details.
 #
-
-# TODO: rename this module
-from .dump_model import *
