+++ python/revng/cli/__init__.py
@@ -1,0 +1,3 @@
+#
+# This file is distributed under the MIT License. See LICENSE.md for details.
+#
