+++ python/revng/cli/merge_dynamic/__init__.py
