+++ python/revng/cli/merge_dynamic/merge_dynamic.py
