+++ python/revng/cli/merge_dynamic/util.py
