+++ python/revng/cli/translate/pipelines/isolate-translate.yml
