+++ python/revng/cli/translate/pipelines/translate.yml
