+++ python/scripts/revng-model-compare
