+++ python/scripts/revng-model-dump
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+#
+# This file is distributed under the MIT License. See LICENSE.md for details.
+#
+from revng.cli.model_dump.__main__ import main
+
+
+if __name__ == "__main__":
+    main()
