+++ python/scripts/revng-model-to-json
@@ -2,7 +2,7 @@
 
 import sys
 
-from revng_dump_model.__main__ import main
+from revng.cli.model_dump.__main__ import main
 
 if __name__ == "__main__":
     try:
