+++ python/setup.py
@@ -1,0 +1,24 @@
+#!/usr/bin/env python3
+
+from setuptools import setup, find_namespace_packages
+
+
+setup(
+    name="revng",
+    version="0.1.0",
+    description="rev.ng Python libraries",
+    author="Filippo Cremonese (rev.ng Labs)",
+    author_email="filippocremonese@rev.ng",
+    url="https://github.com/revng/revng",
+    packages=find_namespace_packages(),
+    include_package_data=True,
+    install_requires=open("requirements.txt").readlines(),
+    scripts=[
+        "scripts/revng",
+        "scripts/revng-merge-dynamic",
+        "scripts/revng-model-compare",
+        "scripts/revng-model-dump",
+        "scripts/revng-model-to-json",
+        "scripts/revng-translate",
+    ],
+)
