--- scripts/revng_support/__init__.py
@@ -1,9 +1,0 @@
-from .support import (
-    shlex_join,
-    log_error,
-    relative,
-    run,
-    get_command,
-    set_verbose,
-    build_command_with_loads,
-)
