--- scripts/revng_translate/__init__.py
@@ -1,1 +1,0 @@
-from .translate import run_translate, register_translate
